pyproject.toml
[project]
name = "chocolate_app"
version = "10.0.2-alpha"
description = ""
authors = [
    { name = "Imprevisible", email = "impr.visible@gmail.com" },
    { name = "Impre-visible", email = "impr.visible@gmail.com" },
]
dependencies = [
    "flask==3.0.3",
    "flask-cors==5.0.0",
    "flask-login==0.6.3",
    "flask-sqlalchemy==3.1.1",
    "get-video-properties==0.1.1",
    "guessit==3.7.1",
    "opencv-python==4.5.5.64",
    "Pillow==9.5.0",
    "pyarr==5.0.0",
    "pycountry==22.3.5",
    "PyPDF2==3.0.1",
    "pypresence==4.2.1",
    "rarfile==4.0",
    "sqlalchemy==2.0.32",
    "Unidecode==1.3.6",
    "Werkzeug==3.0.4",
    "deezer-python>=6.1.0",
    "tinytag>=1.9.0",
    "argparse>=1.4.0",
    "flask-migrate>=4.0.4",
    "tmdbv3api>=1.9.0",
    "natsort>=8.4.0",
    "PyMuPdf>=1.22.5",
    "ffmpeg>=1.4",
    "pandas>=2.2.1",
    "PyJWT>=2.9.0",
    "m3u-parser>=0.4.1",
    "jinja2==3.1.4",
    "greenlet>=3.1.1",
    "PyYaml>=6.0.2",
    "Python-Levenshtein>=0.26.0",
    "schedule>=1.2.2",
    "watchdog>=5.0.3",
    "iso639-lang>=2.5.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "GPL-3.0" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
lint = ["black>=23.7.0"]

[project.scripts]
chocolate_app = "chocolate_app.__main__:start_chocolate"

[tool.pdm.scripts]
chocolate_app = "chocolate_app.__main__:start_chocolate"
